Move external quantizer to separate file back
