+++ nncf/torch/external_hook.py
@@ -12,11 +12,8 @@
 from typing import Any
 
 from nncf.torch.dynamic_graph.context import TracingContext
-from nncf.torch.quantization.debug_interface import QuantizationDebugInterface
 
 EXTERNAL_OP_STORAGE_NAME = "external_op"
-EXTERNAL_QUANTIZERS_STORAGE_NAME = "external_quantizers"
-EXTERNAL_QUANTIZERS_STORAGE_PREFIX = "_nncf." + EXTERNAL_QUANTIZERS_STORAGE_NAME
 
 
 class ExternalOpCallHook:
@@ -43,24 +40,3 @@ class ExternalOpCallHook:
         replica = self._compressed_context.base_module_thread_local_replica
         storage = getattr(replica.nncf, self._storage_name)
         return storage[self._storage_key](*args, **kwargs)
-
-
-class ExternalQuantizerCallHook(ExternalOpCallHook):
-    """
-    External hook which is using quantization storage name and
-    could be constructed with a debug interface.
-    """
-
-    def __init__(
-        self,
-        context: TracingContext,
-        quantizer_storage_key: str,
-        debug_interface: QuantizationDebugInterface = None,
-    ):
-        super().__init__(EXTERNAL_QUANTIZERS_STORAGE_NAME, context, quantizer_storage_key)
-        self.debug_interface = debug_interface
-
-    def __call__(self, *args, **kwargs):
-        if self.debug_interface is not None:
-            self.debug_interface.register_activation_quantize_call(str(self._storage_key))
-        return super().__call__(*args, **kwargs)
