+++ tests/torch/quantization/test_range_init.py
@@ -37,9 +37,9 @@ from nncf.config import NNCFConfig
 from nncf.config.structures import QuantizationRangeInitArgs
 from nncf.torch import utils
 from nncf.torch.checkpoint_loading import load_state
-from nncf.torch.external_hooks import EXTERNAL_QUANTIZERS_STORAGE_NAME
 from nncf.torch.initialization import DefaultInitializingDataLoader
 from nncf.torch.initialization import wrap_dataloader_for_init
+from nncf.torch.quantization.external_quantizer import EXTERNAL_QUANTIZERS_STORAGE_NAME
 from nncf.torch.quantization.init_range import PTRangeInitCollectorParams
 from nncf.torch.quantization.init_range import PTRangeInitParams
 from nncf.torch.quantization.init_range import StatCollectorGenerator
