+++ tests/torch/test_load_model_state.py
@@ -23,9 +23,9 @@ from nncf.torch.checkpoint_loading import KeyMatcher
 from nncf.torch.checkpoint_loading import ProcessedKeys
 from nncf.torch.checkpoint_loading import ProcessedKeyStatus
 from nncf.torch.checkpoint_loading import load_state
-from nncf.torch.external_hooks import EXTERNAL_QUANTIZERS_STORAGE_PREFIX
 from nncf.torch.layers import NNCF_PADDING_VALUE_ATTR_NAME
 from nncf.torch.nncf_module_replacement import replace_modules_by_nncf_modules
+from nncf.torch.quantization.external_quantizer import EXTERNAL_QUANTIZERS_STORAGE_PREFIX
 from tests.torch.helpers import BasicConvTestModel
 from tests.torch.helpers import PTTensorListComparator
